refactor(model-helper): split model(weights) helper functions from federated_main
